Moved the class to a common place
