+++ neighsyncd/neighsyncd.cpp
@@ -1,4 +1,3 @@
-#include <time.h>
 #include <iostream>
 #include "logger.h"
 #include "select.h"
